rename util.ts to util/amounts.ts and factor in suggestions.
